Recreate Card (now PlantCard) using react-bootstrap
